Massive Hiearchy Rework
Reduced redundancy amongst components and increased modularity of components by doing so.

In the process, reworked folder hierarchy to clean and better tell which folder contains what.
