--- operationportal/lib/FilterCheckedIn.dart
@@ -1,4 +1,0 @@
-class FilterCheckedIn
-{
-
-}
