+++ operationportal/lib/Structs/Volunteer.dart
@@ -3,20 +3,33 @@ import 'Class.dart';
 import 'Profile.dart';
 
 class Volunteer {
+  /*“volunteer”: {  “training”: [{“name”: “string”}], “referral”: “string”, “languages”: [“language”: “string”], “newsletter”: “bool”, “contactWhenShort”: “bool”, “phone”: “string”, “email”: “string”,“blueShirt”:bool, “nametag”:bool, “personalInterviewCompleted”:bool, “backgroundCheck”:bool, “yearStarted”:int, “canEditInventory”:bool, “picture”:byte[], “birthday”:DateTime}*/
+  int id;
   String firstName;
   String lastName;
+  String preferredName;
   Class mClass;
   Bus mBus;
 
+  String affiliation;
+  String referral;
+  bool orientation;
 
-  Volunteer({this.firstName, this.lastName, this.mClass, this.mBus});
+
+  Volunteer({this.id, this.firstName, this.lastName, this.preferredName,
+    this.mClass, this.mBus, this.affiliation, this.referral, this.orientation});
 
   factory Volunteer.fromJson(Map<String, dynamic> json) {
     return Volunteer(
+      id: json['id'],
       firstName: json['firstName'],
       lastName: json['lastName'],
-      mClass: Class.fromJson(json['class']),
-      mBus: Bus.fromJson(json['bus']),
+      preferredName: json['preferredName'],
+      mClass: json['class'] != null ? Class.fromJson(json['class']) : null,
+      mBus: json['bus'] != null ? Bus.fromJson(json['bus']) : null,
+      affiliation: json['affiliation'],
+      referral: json['referral'],
+      orientation: json['orientation'],
     );
   }
 }
