--- operationportal/lib/Volunteer/Volunteer_InactiveDashboard.dart
@@ -1,43 +1,0 @@
-import 'package:operationportal/BuildPresets/AppBar.dart';
-import 'package:operationportal/BuildPresets/InactiveDashboard.dart';
-import 'package:operationportal/Login.dart';
-import 'package:operationportal/Structs/Profile.dart';
-import 'package:flutter/material.dart';
-
-class Volunteer_InactiveDashboard_Page extends StatefulWidget {
-  Volunteer_InactiveDashboard_Page({Key key, this.profile, this.accessToken}) : super(key: key);
-
-  final Profile profile;
-  final String accessToken;
-
-  @override
-  Volunteer_InactiveDashboard_State createState() => Volunteer_InactiveDashboard_State();
-}
-
-class Volunteer_InactiveDashboard_State extends State<Volunteer_InactiveDashboard_Page>
-{
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(widget.profile.role),
-        actions: <Widget>[
-          buildLogoutButton(context),
-        ],
-      ),
-      resizeToAvoidBottomPadding: false,
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.start,
-          crossAxisAlignment: CrossAxisAlignment.center,
-          children: <Widget>[
-            buildPictureNameRow(widget.profile.firstName, widget.profile.lastName),
-            buildQRButton(widget.accessToken, context),
-            buildNotice(),
-          ],
-        ),
-      ),
-    );
-  }
-}
